requirements.txt
PyGithub
west>=0.8.0
